+++ sdk/typespec/typespec_client_core/src/http/policies/retry/mod.rs
@@ -5,9 +5,9 @@ mod exponential;
 mod fixed;
 mod none;
 
-pub use exponential::*;
-pub use fixed::*;
-pub use none::*;
+pub(crate) use exponential::*;
+pub(crate) use fixed::*;
+pub(crate) use none::*;
 
 use crate::{
     http::{